requirements.txt
audioop-lts==0.2.1
py-cord==2.6.1
python-dotenv==1.0.1